Implement RequestCounter
RequestCounter is a datastore to keep track the number of request coming from an IP into an API endpoint.
The actual datastore is Rails.cache using MemoryStore (see commit fc71d0b).
MemoryStore is chosen because:
- It will be faster than a database backend for example as it is in memory
- It comes with Rails - no need to install another software such as memcached

The drawback of this datastore is the data will not be retained when Rails restart (which for the purpose of API request limiting is acceptable).