Refactor - extract RequestCounterDataStore from RequestCounter
The refactor is to clarify the intent of RequestCounter which is an entry in the datastore.

The extracted RequestCounterDataStore is a singleton class that has the reference to the actual data store
used to hold RequestCounter's data - in this case we are using MemoryStore.

By extracting RequestCounterDataStore, we also gained:

* A flexiblity of changing the data store later.
* We decouple the data store requirement from Rails.cache